Try to fix MDEC decoding

+++ psx/mdec.h
@@ -102,7 +102,7 @@ private:
     void decode_collected_blocks();
 
     static int16_t sign_extend(uint16_t value);
-    static int16_t clamp(int16_t value);
+    static int16_t clamp(int32_t value);
     // Reads the next block from the input queue into the provided buffer
     bool read_next_block(std::vector<uint16_t>& block);
     // Decode the RLE-encoded block
@@ -117,7 +117,7 @@ private:
     bool decode_next_block(const std::vector<uint8_t>& quant, std::vector<int16_t>& buffer);
     static uint8_t index(uint8_t i, uint8_t j) { return i * 8 + j; }
     void idct(std::vector<int32_t>& result, std::vector<int16_t>& block);
-    static int16_t clamp_color(int16_t value);
+    static int16_t clamp_color(int32_t value);
     void yuv_to_rgb(std::vector<uint8_t>& r, std::vector<uint8_t>& g, std::vector<uint8_t>& b,
                     const std::vector<int32_t>& cr, const std::vector<int32_t>& cb, const std::vector<int32_t>& y,
                     uint8_t x_offset, uint8_t y_offset);
